Make crawl slower
I have observed that for lines that have more text, or for those lines
that are just separated with a <br> (and therefore have almost no space
between them), there is not enough time to read. The user has to try to
read too fast, which spoils the experience. This commit increases the
time it takes for the scroll/crawl to happen in 0.3 seconds. Seems to be
enough, although not so slow to the point of becoming boring.

Sadly, I cannot update the timing on the video on youtube, although the
updated animation can bee seen on my blog.

• https://fernandobasso.dev/web-development/math-poem-star-wars.html
• https://www.youtube.com/watch?v=RiL1uJ63CoM